# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Warehouse:
    type: object
host: localhost:8080
info:
  contact:
    email: support@b-resto.com
    name: Soporte API
  description: API RESTful para sistema POS de restaurantes
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: B-Resto API
  version: "1.0"
paths:
  /admin/resources:
    post:
      consumes:
      - application/json
      description: Crea un nuevo recurso (solo admin)
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Resource created by admin'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear recurso
      tags:
      - resources
  /admin/resources/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un recurso (solo admin)
      parameters:
      - description: ID del recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Resource deleted by admin'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar recurso
      tags:
      - resources
  /admin/users:
    get:
      consumes:
      - application/json
      description: Obtiene lista de todos los usuarios (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: 'users: array de usuarios'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: Failed to fetch users'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar usuarios
      tags:
      - users
...
  /profile:
    get:
      consumes:
      - application/json
      description: Obtiene el perfil del usuario autenticado desde el token JWT
      produces:
      - application/json
      responses:
        "200":
          description: 'user: datos del usuario'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid token claims'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener perfil del usuario
      tags:
      - users
...
  /resources/{id}:
    put:
      consumes:
      - application/json
      description: Actualiza un recurso (endpoint de ejemplo)
      parameters:
      - description: ID del recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Resource updated'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar recurso
      tags:
      - resources
...
  /warehouses:
    get:
      consumes:
      - application/json
      description: Obtiene lista de todos los almacenes
      parameters:
      - description: Filtrar por estado activo
        enum:
        - "true"
        - "false"
        in: query
        name: is_active
        type: string
      - description: Filtrar por compañía/sucursal
        in: query
        name: company_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de warehouses'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar almacenes
      tags:
      - warehouses
    post:
      consumes:
      - application/json
      description: Crea un nuevo almacén
      parameters:
      - description: Datos del almacén
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/models.Warehouse'
      produces:
      - application/json
      responses:
        "201":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 'error: código duplicado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear almacén
      tags:
      - warehouses
  /warehouses/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un almacén (soft delete)
      parameters:
      - description: ID del almacén
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Warehouse deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Warehouse not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar almacén
      tags:
      - warehouses
    get:
      consumes:
      - application/json
      description: Obtiene un almacén por ID
      parameters:
      - description: ID del almacén
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: warehouse'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Warehouse not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener almacén
      tags:
      - warehouses
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un almacén existente
      parameters:
      - description: ID del almacén
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/models.Warehouse'
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Warehouse not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar almacén
      tags:
      - warehouses
  /warehouses/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Cambia el estado is_active de un almacén
      parameters:
      - description: ID del almacén
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Warehouse not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Activar/Desactivar almacén
      tags:
      - warehouses
securityDefinitions:
  Bearer:
    description: Escribir "Bearer {token}" para autenticación JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"